Move all mocks under new /tests/mocks dir
