+++ .trunk/configs/bandit.yaml
@@ -1,1 +1,1 @@
-exclude_dirs: ["unit_tests", "integration_tests"]
+exclude_dirs: ["unit_tests", "integration_tests", "tests"]
