+++ tests/mocks/repos/__init__.py
@@ -1,6 +1,6 @@
 from pytest import MonkeyPatch
 
-import tests.unit_tests.mocks.repos.family_repo as mock_repo  # type: ignore
+import tests.mocks.repos.family_repo as mock_repo  # type: ignore
 from app.repository.protocols import FamilyRepo
 
 mock_repo: FamilyRepo
