+++ tests/mocks/repos/bad_collection_repo.py
