+++ tests/mocks/repos/bad_event_repo.py
