+++ tests/mocks/services/analytics_service.py
