+++ tests/unit_tests/conftest.py
@@ -33,24 +33,24 @@ from app.repository import (
     metadata_repo,
     organisation_repo,
 )
-from tests.unit_tests.mocks.repos import create_mock_family_repo
-from tests.unit_tests.mocks.repos.app_user_repo import mock_app_user_repo
-from tests.unit_tests.mocks.repos.collection_repo import mock_collection_repo
-from tests.unit_tests.mocks.repos.config_repo import mock_config_repo
-from tests.unit_tests.mocks.repos.corpus_repo import mock_corpus_repo
-from tests.unit_tests.mocks.repos.document_repo import mock_document_repo
-from tests.unit_tests.mocks.repos.event_repo import mock_event_repo
-from tests.unit_tests.mocks.repos.geography_repo import mock_geography_repo
-from tests.unit_tests.mocks.repos.metadata_repo import mock_metadata_repo
-from tests.unit_tests.mocks.repos.organisation_repo import mock_organisation_repo
-from tests.unit_tests.mocks.services.analytics_service import mock_analytics_service
-from tests.unit_tests.mocks.services.app_user_service import mock_app_user_service
-from tests.unit_tests.mocks.services.collection_service import mock_collection_service
-from tests.unit_tests.mocks.services.config_service import mock_config_service
-from tests.unit_tests.mocks.services.corpus_service import mock_corpus_service
-from tests.unit_tests.mocks.services.document_service import mock_document_service
-from tests.unit_tests.mocks.services.event_service import mock_event_service
-from tests.unit_tests.mocks.services.family_service import mock_family_service
+from tests.mocks.repos import create_mock_family_repo
+from tests.mocks.repos.app_user_repo import mock_app_user_repo
+from tests.mocks.repos.collection_repo import mock_collection_repo
+from tests.mocks.repos.config_repo import mock_config_repo
+from tests.mocks.repos.corpus_repo import mock_corpus_repo
+from tests.mocks.repos.document_repo import mock_document_repo
+from tests.mocks.repos.event_repo import mock_event_repo
+from tests.mocks.repos.geography_repo import mock_geography_repo
+from tests.mocks.repos.metadata_repo import mock_metadata_repo
+from tests.mocks.repos.organisation_repo import mock_organisation_repo
+from tests.mocks.services.analytics_service import mock_analytics_service
+from tests.mocks.services.app_user_service import mock_app_user_service
+from tests.mocks.services.collection_service import mock_collection_service
+from tests.mocks.services.config_service import mock_config_service
+from tests.mocks.services.corpus_service import mock_corpus_service
+from tests.mocks.services.document_service import mock_document_service
+from tests.mocks.services.event_service import mock_event_service
+from tests.mocks.services.family_service import mock_family_service
 
 
 @pytest.fixture
