+++ tests/unit_tests/service/test_collection_service.py
@@ -8,9 +8,7 @@ from app.model.collection import (
     CollectionWriteDTO,
 )
 from tests.helpers.collection import create_collection_write_dto
-from tests.unit_tests.mocks.repos.collection_repo import (
-    create_collection_read_dto as create_dto,
-)
+from tests.mocks.repos.collection_repo import create_collection_read_dto as create_dto
 
 
 def _to_write_dto(dto: CollectionReadDTO) -> CollectionWriteDTO:
